requirements.txt
requests
beautifulsoup4
lxml
pydantic
pydantic-settings
sqlalchemy
httpx
tenacity
python-dotenv
Pillow
numpy
diffusers
transformers
torch
feedparser
fastapi
uvicorn
jinja2
